+++ src/d/actor/d_a_npc_henna.cpp
@@ -379,7 +379,7 @@ static void henna_shop(npc_henna_class* i_this) {
     }
     cLib_addCalcAngleS2(&i_this->actor.current.angle.y, i_this->field_0x618, 2, unkInt1);
     /* dSv_event_flag_c::F_0461 - Fishing Pond - First time entered fishing house */
-    if ((!dComIfGs_isEventBit((u16)dSv_event_flag_c::saveBitLabels[461]) || data_80450C9C != 0) &&
+    if ((!dComIfGs_isEventBit(dSv_event_flag_c::saveBitLabels[461]) || data_80450C9C != 0) &&
         i_this->field_0x7d7 == 0 && i_this->field_0x6a0 == 56 + TREG_S(7))
     {
         dComIfGp_event_reset();
@@ -388,7 +388,7 @@ static void henna_shop(npc_henna_class* i_this) {
             i_this->cam_mode = 60;
         } else {
             /* dSv_event_flag_c::F_0461 - Fishing Pond - First time entered fishing house */
-            dComIfGs_onEventBit((u16)dSv_event_flag_c::saveBitLabels[461]);
+            dComIfGs_onEventBit(dSv_event_flag_c::saveBitLabels[461]);
             i_this->cam_mode = 1;
         }
         i_this->field_0x7d7 = 1;
@@ -890,9 +890,9 @@ static void demo_camera(npc_henna_class* i_this) {
                 i_this->cam_mode = 0;
             }
             /* dSv_event_flag_c::F_0463 - Fishing Pond - Reserved for fishing */
-            dComIfGs_offEventBit((u16)dSv_event_flag_c::saveBitLabels[0x1cf]);
+            dComIfGs_offEventBit(dSv_event_flag_c::saveBitLabels[0x1cf]);
             /* dSv_event_flag_c::F_0464 - Fishing Pond - Reserved for fishing */
-            dComIfGs_offEventBit((u16)dSv_event_flag_c::saveBitLabels[0x1d0]);
+            dComIfGs_offEventBit(dSv_event_flag_c::saveBitLabels[0x1d0]);
         }
     }
 
@@ -994,7 +994,7 @@ static void demo_camera_shop(npc_henna_class* i_this) {
 
     if (!dComIfGp_event_runCheck()
             /* dSv_event_flag_c::F_0465 - Fishing Pond - Reserved for fishing */
-        && !dComIfGs_isEventBit((u16)dSv_event_flag_c::saveBitLabels[0x1d1]))
+        && !dComIfGs_isEventBit(dSv_event_flag_c::saveBitLabels[0x1d1]))
     {
         i_this->field_0x756++;
         if (i_this->field_0x756 < 200) {
@@ -2728,9 +2728,9 @@ static int daNpc_Henna_Create(fopAc_ac_c* i_this) {
                 }
             }
             /* dSv_event_flag_c::F_0463 - Fishing Pond - Reserved for fishing */
-            dComIfGs_offEventBit((u16)dSv_event_flag_c::saveBitLabels[0x1cf]);
+            dComIfGs_offEventBit(dSv_event_flag_c::saveBitLabels[0x1cf]);
             /* dSv_event_flag_c::F_0464 - Fishing Pond - Reserved for fishing */
-            dComIfGs_offEventBit((u16)dSv_event_flag_c::saveBitLabels[0x1d0]);
+            dComIfGs_offEventBit(dSv_event_flag_c::saveBitLabels[0x1d0]);
             a_this->field_0x618 = i_this->shape_angle.y = i_this->current.angle.y = newAngle;
             lbl_82_bss_91 = dComIfGs_getEventReg(0xf63f);
             daNpc_Henna_Execute(a_this);
@@ -2775,7 +2775,7 @@ static int daNpc_Henna_Create(fopAc_ac_c* i_this) {
                               NULL, -1);
                 data_80450C99 = 0;
                 /* dSv_event_flag_c::F_0463 - Fishing Pond - Reserved for fishing */
-                if (!dComIfGs_isEventBit((u16)dSv_event_flag_c::saveBitLabels[0x1cf])) {
+                if (!dComIfGs_isEventBit(dSv_event_flag_c::saveBitLabels[0x1cf])) {
                     a_this->field_0x70a = 4;
                     a_this->field_0x734 = 1;
                 } else {
